+++ backend/src/main/java/com/company/flowmodoro/features/session/dtos/SessionUpdateDTO.java
@@ -11,9 +11,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class SessionUpdateDTO {
   private Long id;
-  private Double focus;
+  private Long focus;
   private Double ratio;
-  private Double rest;
+  private Long rest;
   private Integer interruptions;
   private Long task;
 }
